assert file existence

+++ src/captcha/index.ts
@@ -5,7 +5,7 @@ import {promisify} from 'util';
 import {exec} from 'child_process';
 import {ICaptchaResponse} from './interface';
 import {isValidLanguage} from '../language/';
-import {isNumberString} from '../util';
+import {isNumberString, readAsString} from '../util';
 import {InvalidCaptchaParameterError} from './error';
 
 const runCommand = promisify(exec);
@@ -22,9 +22,12 @@ export const createCaptcha = async (
 	speed?: string,
 	gap?: string,
 ): Promise<ICaptchaResponse> => {
+	const id = uuid();
 	const language = lang ?? 'en-us';
 	const sp = speed ?? '100';
 	const gp = gap ?? '10';
+	console.log(`generating captcha: ${id} (lang: ${language}, speed: ${sp}, gap:${gp})`);
+
 	if (!isValidLanguage(language)) {
 		throw new InvalidCaptchaParameterError('language', language);
 	}
@@ -37,8 +40,6 @@ export const createCaptcha = async (
 		throw new InvalidCaptchaParameterError('gap', gp);
 	}
 
-	const id = uuid();
-	console.log(`generating captcha: ${id} (lang: ${language}, speed: ${sp}, gap:${gp})`);
 	const cap = captcha.create();
 	const {
 		data: svgData,
@@ -50,10 +51,12 @@ export const createCaptcha = async (
 	await runCommand(
 		`espeak-ng "${withSpaces}" -w ${filename} -v ${language} -s ${sp} -g ${gp}`,
 	);
-	// TO-DO: check for error
-	const fileData = fs.readFileSync(filename, {
-		encoding: 'base64',
-	});
+
+	if (!fs.existsSync(filename)) {
+		throw new Error('failed to create captcha audio');
+	}
+
+	const fileData = readAsString(filename);
 	fs.removeSync(filename);
 	return {
 		id,
